feature: auth

+++ WebApi/Models/Location.cs
@@ -1,10 +1,12 @@
 using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace WebApi.Models
 {
     public class Location
     {
-        public int LocationId { get; set; }
+        public long LocationId { get; set; }
         public decimal Latitude { get; set; }
         public decimal Longitude { get; set; }
         public DateTime Timestamp { get; set; }
